Save and load HUD layout, prefer raster font on low resolution, select vector font size close to raster one, make OSDCLASS::scale floating point

+++ desmume/src/frontend/modules/osd/agg/agg_osd.h
@@ -112,7 +112,7 @@ public:
 	char	name[7];		// for debuging
 	bool    singleScreen;
 	bool    swapScreens;
-	int		scale;
+	double	scale;
 #ifdef AGG2D_USE_VECTORFONTS
 	bool	useVectorFonts;
 #endif
